Reorganize API file into multiple separate files
