+++ server/core/views/api/accounts.py
@@ -1,0 +1,37 @@
+from core.serializers import StudentSerializer, UserSerializer
+from rest_framework import permissions, status
+from rest_framework.decorators import api_view
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+# Login API calls
+
+
+@api_view(["GET"])
+def current_user(request):
+    """
+    Determine the current user by their token, and return their data
+    Permissions: authenticated (automatic)
+    """
+
+    serializer = UserSerializer(request.user)
+    return Response(serializer.data)
+
+
+class StudentAccount(APIView):
+    """
+    Create a new student and return the access/refresh token pair.
+    This is used during the signup process.
+    Errors are not being handled yet.
+    Permissions: any (don't have permissions before user is created)
+    """
+
+    permission_classes = (permissions.AllowAny,)
+    authentication_classes = []
+
+    def post(self, request, format=None):
+        serializer = StudentSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
