+++ server/core/views/api/studentreg.py
@@ -1,103 +1,13 @@
 import core.permissions as custom_permissions
 from core.models import Program, StudentRegistration
-from core.serializers import (
-    ClassSerializer,
-    ProgramSerializer,
-    StudentRegSerializer,
-    StudentSerializer,
-    UserSerializer,
-)
+from core.serializers import StudentRegSerializer
 from django.db import transaction
 from django.forms.models import model_to_dict
-from rest_framework import permissions, status, viewsets
-from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-# Login API calls
-
-
-@api_view(["GET"])
-def current_user(request):
-    """
-    Determine the current user by their token, and return their data
-    Permissions: authenticated (automatic)
-    """
-
-    serializer = UserSerializer(request.user)
-    return Response(serializer.data)
-
-
-class StudentAccount(APIView):
-    """
-    Create a new student and return the access/refresh token pair.
-    This is used during the signup process.
-    Errors are not being handled yet.
-    Permissions: any (don't have permissions before user is created)
-    """
-
-    permission_classes = (permissions.AllowAny,)
-    authentication_classes = []
-
-    def post(self, request, format=None):
-        serializer = StudentSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-# Dashboard API calls
-
-
-class TeacherProgramViewSet(viewsets.ReadOnlyModelViewSet):
-    """
-    API endpoint that allows programs with open teacherreg to be viewed.
-    Permissions: authenticated+teacher
-    """
-
-    permission_classes = (custom_permissions.IsTeacher,)
-    # TODO(mvadari): let's move this to a get_active_programs() in a Program Manager
-    queryset = Program.objects.filter(teacher_reg_open=True).order_by("edition", "name")
-    serializer_class = ProgramSerializer
-
-
-class ClassCatalog(APIView):
-    """
-    API endpoint that returns all classes.
-    Permissions: authenticated
-    # TODO have some publicly accessible version of this API
-    """
-
-    def get(self, request, program, edition, format=None):
-        classes = Program.objects.get(name=program, edition=edition).classes.all()
-        return Response([ClassSerializer(clazz).data for clazz in classes])
-
-
-@api_view(["GET"])
-@permission_classes([custom_permissions.IsStudent])
-def get_student_dashboard(request):
-    user = request.user
-
-    previous_programs = StudentRegistration.get_previous_programs(user)
-    previous_json = [{"name": str(p), "url": p.url} for p in previous_programs]
-
-    current_programs = StudentRegistration.get_current_programs(user)
-
-    open_programs = Program.get_open_student_programs().exclude(id__in=current_programs)
-    # TODO add grade checks
-
-    current_json = [{"name": str(p), "url": p.url, "registered": False} for p in open_programs] + [
-        {"name": str(p), "url": p.url, "registered": True} for p in current_programs
-    ]
-    current_json.sort(key=lambda p: p["name"])  # TODO edit to be better
-
-    # TODO clean up API so we don't need weird conditionals in the UI
-
-    return Response({"previous": previous_json, "current": current_json})
-
-
 # Registration API calls
+
 class StudentRegAPI(APIView):
     """
     Determine the current studentreg object by the user and the program/edition
@@ -115,6 +25,33 @@ class StudentRegAPI(APIView):
         studentreg, _ = StudentRegistration.objects.get_or_create(student=user, program=prog)
         # TODO figure out how to change the default regstatus based on the program's status
         return Response(StudentRegSerializer(studentreg).data)
+
+
+class StudentProgramClasses(APIView):
+    """
+    Get/update a student's class list for a given program.
+    Permissions: logged in, is student, has studentreg object
+    """
+
+    def get(self, request, program, edition, format=None):
+        # user = request.user
+
+        # # TODO make this check better
+        # prog = Program.objects.filter(name=program, edition=edition)[0]
+        # studentreg = StudentRegistration.objects.get(student=user, program=prog)
+        # classes = studentreg.get_classes()
+        # timeslots = Timeslot.objects.filter(program=prog)
+
+        ret = {
+            "classes": [
+                "How to beat your dad at chess, MIT style",
+                "From Neurons to Thoughts: An Introduction to the Human Mind and Brain",
+                None,
+            ],
+            "timeslots": ["Sat 10-11", "Sat 11-12", "Sat 12-1"],
+        }
+
+        return Response(ret)
 
 
 class Profile(APIView):
@@ -306,33 +243,3 @@ class Availability(APIView):
         studentreg = StudentRegistration.objects.get(student=user, program=program)
 
         return studentreg
-
-
-class StudentProgramClasses(APIView):
-    """
-    Get/update a student's class list for a given program.
-    Permissions: logged in, is student, has studentreg object
-    """
-
-    def get(self, request, program, edition, format=None):
-        # user = request.user
-
-        # # TODO make this check better
-        # prog = Program.objects.filter(name=program, edition=edition)[0]
-        # studentreg = StudentRegistration.objects.get(student=user, program=prog)
-        # classes = studentreg.get_classes()
-        # timeslots = Timeslot.objects.filter(program=prog)
-
-        ret = {
-            "classes": [
-                "How to beat your dad at chess, MIT style",
-                "From Neurons to Thoughts: An Introduction to the Human Mind and Brain",
-                None,
-            ],
-            "timeslots": ["Sat 10-11", "Sat 11-12", "Sat 12-1"],
-        }
-
-        return Response(ret)
-
-
-# TODO figure out object permissions
